fix: fix sticky tabs + update CR notes
