+++ superset-frontend/spec/javascripts/explore/components/EmbedCodeButton_spec.jsx
@@ -27,6 +27,7 @@ import fetchMock from 'fetch-mock';
 import EmbedCodeButton from 'src/explore/components/EmbedCodeButton';
 import * as exploreUtils from 'src/explore/exploreUtils';
 import * as common from 'src/utils/common';
+import * as urlUtils from 'src/utils/urlUtils';
 
 const ENDPOINT = 'glob:*/r/shortner/';
 
@@ -53,7 +54,7 @@ describe('EmbedCodeButton', () => {
 
   it('should create a short, standalone, explore url', () => {
     const spy1 = sinon.spy(exploreUtils, 'getExploreLongUrl');
-    const spy2 = sinon.spy(common, 'getShortUrl');
+    const spy2 = sinon.spy(urlUtils, 'getShortUrl');
 
     const wrapper = mount(
       <ThemeProvider theme={supersetTheme}>
