+++ superset-frontend/src/dashboard/stylesheets/builder.less
@@ -22,10 +22,6 @@
   flex-grow: 1;
   display: flex;
   flex-direction: column;
-
-  &--no-header {
-    border-top: 1px solid @gray-light;
-  }
 }
 
 /* only top-level tabs have popover, give it more padding to match header + tabs */
